quotes-api/pyproject.toml
[tool.poetry]
name = "quotes-api"
version = "0.1.0"
description = "API that returns random quotes"
authors = ["Andrew Guest <andrew@aguest.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
secure = "^0.3.0"
gunicorn = "^23.0.0"
uvicorn-worker = "^0.2.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
